requirements.txt
# RAFT-DVC Dependencies
# This project was developed and tested on NVIDIA RTX 5090 (24GB VRAM)
# Recommended: CUDA-capable GPU with at least 16GB VRAM for training 128³ volumes

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=1.21.0,<2.0.0
scipy>=1.7.0

# Data Processing
pyyaml>=5.4
h5py>=3.0.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Progress Tracking
tqdm>=4.60.0

# Logging and Monitoring
tensorboard>=2.8.0

# Optional: For memory profiling
# psutil>=5.8.0

# Optional: For 3D visualization
# mayavi>=4.7.0
# vtk>=9.0.0

# Note: For optimal performance on RTX 5090, install PyTorch with CUDA support:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
